+++ packages/faustwp-core/src/components/Toolbar/index.ts
@@ -1,5 +1,6 @@
 export { Toolbar, FaustToolbarNodes, FaustToolbarContext } from './Toolbar.js';
 export { ToolbarNode } from './ToolbarNode.js';
+export { ToolbarNodeSkeleton } from './ToolbarNodeSkeleton.js';
 export { ToolbarItem } from './ToolbarItem.js';
 export { ToolbarSubmenu } from './ToolbarSubmenu.js';
 export { ToolbarSubmenuWrapper } from './ToolbarSubmenuWrapper.js';