+++ packages/faustwp-core/src/getWordPressProps.tsx
@@ -2,6 +2,7 @@ import { GetServerSidePropsContext, GetStaticPropsContext } from 'next';
 import type { DocumentNode } from 'graphql';
 import { SeedNode, SEED_QUERY } from './queries/seedQuery.js';
 import { getPossibleTemplates, getTemplate } from './getTemplate.js';
+import { FaustTemplateProps } from './components/WordPressTemplate.js';
 import { addApolloState, getApolloClient } from './client.js';
 import { getConfig } from './config/index.js';
 import { hooks } from './wpHooks/index.js';
@@ -23,13 +24,30 @@ export type WordPressTemplate = React.FC & {
   ) => { [key: string]: any };
 };
 
+export interface FaustTemplate<Data>
+  extends React.FC<FaustTemplateProps<Data>> {
+  query?: WordPressTemplate['query'];
+  variables?: WordPressTemplate['variables'];
+}
+
 export interface GetWordPressPropsConfig<Props = Record<string, unknown>> {
   ctx: GetServerSidePropsContext | GetStaticPropsContext;
   props?: Props;
   revalidate?: number | boolean;
 }
-
-export async function getWordPressProps(options: GetWordPressPropsConfig) {
+export async function getWordPressProps(
+  options: GetWordPressPropsConfig,
+): Promise<
+  | {
+      props: {
+        [key: string]: any;
+      };
+      revalidate?: number | boolean | undefined;
+    }
+  | {
+      notFound: true;
+    }
+> {
   const { templates } = getConfig();
 
   if (!templates) {