+++ packages/faustwp-core/tests/components/Toolbar/ToolbarSubmenu.test.tsx
@@ -1,0 +1,11 @@
+/** @jest-environment jsdom */
+import * as React from 'react';
+import '@testing-library/jest-dom';
+import { render, screen } from '@testing-library/react';
+import { ToolbarSubmenu } from '../../../src/components/Toolbar';
+
+test('renders the component correctly', () => {
+  render(<ToolbarSubmenu id="1" />);
+  const aElement = screen.getByRole(/list/i);
+  expect(aElement).toBeInTheDocument();
+});