+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @faustwp/blocks
 
+## 0.2.1
+
+### Patch Changes
+
+- fed67a4: Added: Throw an error if the `blocks` prop was not set on the `<WordPressBlocksViewer>` component
+
+## 0.2.0
+
+### Minor Changes
+
+- 9f4c23d: **[BREAKING]** Renamed `contentBlocks` field to `blocks` in WordpressBlocksViewer
+
+### Patch Changes
+
+- 7be7dfc: Properly exports `cjs` dist build for CommonJS support
+- 1253e86: Bug: Reference correct type definitions in package.json
+- 1c8f79b: Type Definition fixes+improvements for Blocks and Core
+
+## 0.1.1
+
+### Patch Changes
+
+- ef9ebdc: Change DefaultBlock to use `div` instead of `span`
+
 ## 0.1.0
 
 ### Minor Changes